% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cal_nnet_pval}
\alias{cal_nnet_pval}
\title{explore the data by NNet}
\usage{
cal_nnet_pval(nnet_obj)
}
\arguments{
\item{geno:}{the additive geno matrix with each column as a sample and each row as a variant.}

\item{pheno:}{the phenotype dataframe that each row is a sample.
The column include a response column named as response. It is multinomial.While other columns are covariates or variables.}

\item{nnet_obj:}{object of output of nnet::multinom}
}
\value{
two-sided pvalues from zscores (coefficients/standard.error) nnet:: multinom regression.
}
\description{
this function calculate the zscores and calculate the two-sided pvals from the results of nnet.
}
