% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_V2}
\alias{get_V2}
\title{get statistic V}
\usage{
get_V2(X, G, Y, mu)
}
\arguments{
\item{X}{matrix contains all covariates(non-interest items)}

\item{G}{matrix contains all Geno parts GxE parts(interest items to be test)}

\item{Y}{matrix contains interested response.}

\item{mu}{estimated y from  multinomial regression. fitted values of output of nnet::multinom}
}
\value{
V_list a list of V's calculated.
}
\description{
this function calculate the statistics V for each categories. This function is wraped by invariant_pval_cal
In practice, we usually use the Rcpp function get_V for computation speed benefit.
}
