% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{invariant_pval_cal}
\alias{invariant_pval_cal}
\title{this function calculate the core pvalues of invariant. This function is wraped by run_invariant_pval}
\usage{
invariant_pval_cal(X, G, Y, mu, VQstat = FALSE)
}
\arguments{
\item{X}{matrix contains all covariates(non-interest items)}

\item{G}{matrix contains all Geno parts GxE parts(interest items to be test)}

\item{Y}{matrix contains interested response.}

\item{mu}{estimated y from  multinomial regression. fitted values of output of nnet::multinom}

\item{VQstat}{bool, if TRUE, returns the V & Q statistics as well.}
}
\description{
this function calculate the core pvalues of invariant. This function is wraped by run_invariant_pval
}
