% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pca_analysis}
\alias{pca_analysis}
\title{get the top PCs from PCA, Part I}
\usage{
pca_analysis(count_m, subj_id, num_top_PCs = 10)
}
\arguments{
\item{count_m}{a matrix/data frame with each column as genes, each rows as subjects.}

\item{subj_id}{a vector with length as same as rows of count_m, it gives the id of subject of count_m}

\item{num_top_PCs}{the number of top PCs that would be picked.}
}
\value{
pca_res output list of function pca_analysis
}
\description{
this function get the top PCs from PCA
}
