% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_topPCs}
\alias{predict_topPCs}
\title{get the top PCs from PCA ,Part II}
\usage{
predict_topPCs(pca_res, subj_id, count_m)
}
\arguments{
\item{pca_res}{output list of function pca_analysis}

\item{subj_id}{a vector with length as same as rows of count_m, it gives the id of subject of count_m}

\item{count_m}{matrix/data frame with each column as genes, each rows as subjects.}
}
\value{
topPCs matrix with columns as expected columns of PCs.
}
\description{
this function get the top PCs from PCA
}
