% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pval_plot_QQ}
\alias{pval_plot_QQ}
\title{pval_plot_hist
function pval_plot_QQ
plot hist plot and QQ plot for output lists of total_pvals.}
\usage{
pval_plot_QQ(pval_list, main_name)
}
\arguments{
\item{pval_list}{The list with each items are the matrix of pvalues from function cal_nnet_pval}

\item{main_name}{The mainname on the figures.}

\item{flie_name}{The filename}
}
\value{
p save the plots with designed names.
}
\description{
pval_plot_hist
function pval_plot_QQ
plot hist plot and QQ plot for output lists of total_pvals.
}
\examples{
pval_list=total_pval_uv
main_name="hist_pval_uv"
pval_plot_QQ(pval_list=total_pval_uv,
main_name="Distributions of pvalues with environment factor uv")

}
